12.4.Updating an Entity with New Fields{Field Default Value}
Field Default Value
Next, we need to make sure these new fields are populated on our dummy articles. Open ArticleAdminController.
Oh, but first, remember that, in the Article entity, heartCount is required in the database
Actually, to be more clear: nullable=true means that it is allowed to be null in the database. If you don't see nullable, it uses the default value, which is false.
Anyways, this means that heartCount needs a value! But here's a cool idea: once our admin area is fully finished, when an author creates a new article, they shouldn't need to set the heartCount manually. I mean, it's not like we expect the form to have a "heart count" input field on it. Nope, we expect it to automatically default to zero for new articles.
So... how can we give a property a default value in the database? By giving it a default value in PHP: $heartCount = 0