12.5.Updating an Entity with New Fields{Using the new Fields}
Using the new Fields
Ok, back to ArticleAdminController! Add $article->setAuthor() and use the same data we had on the original, hardcoded articles.
Then, ->setHeartCount() and give this a random number between, how about, 5 and 100.
And finally, ->setImageFilename(). The file we've been using is called asteroid.jpeg. Keep using that.
Excelente! Because we already have a bunch of records in the database where these fields are blank, just to keep things simple, let's delete the table entirely and start fresh. Do that with:

php bin/console doctrine:query:sql "TRUNCATE TABLE article"
If you check out the page now and refresh... cool, it's empty. Now, go to /admin/article/new and... refresh a few times. Awesome! Check out the homepage!
We have articles... but actually... this author is still hardcoded in the template. Easy fix!