12.7.Updating an Entity with New Fields{Updating the Image Path}
Updating the Image Path
The last piece that's still hardcoded is this image. Go back to homepage.html.twig. The image path uses asset('images/asteroid.jpeg')
So... this is a little bit tricky, because only part of this - the asteroid.jpeg part - is stored in the database. One solution would be to use Twig's concatenation operator, which is ~, then article.imageFilename
You don't see the ~ much in Twig, but it works like a . in PHP.
That's fine, but a nicer way would be to create a new method that does this for us.