12.8.Updating an Entity with New Fields{Updating the Image Path}
Open Article and, at the bottom, create a new public function getImagePath()
Inside, return images/ and then $this->getImageFilename()
Thanks to this, in the template, we only need to say article.imagePath.
And yea, imagePath is totally not a real property on Article! But thanks to the kung fu powers of Twig, this works fine.
Oh, and side note: notice that there is not an opening slash on these paths.
As a reminder, you do not need to include the opening / when using the asset() function: Symfony will add it there automatically.
Ok, try it out - refresh! It still works! And now that we've centralized that method, in show.html.twig, it's super easy to make the same change: article.imagePath.
Awesome. And when you click on the show page, it works too.
Next! Now that the heart count is stored in the database, let's make our AJAX endpoint that "likes" an article actually work correctly. Right now, it does nothing, and returns a random number. We can do better.