13.1.Updating an Entity In the previous tutorial, we created our heart feature! You click on the heart, it makes an Ajax request back to the server, and returns the new number of hearts. It's all very cute. In theory... when we click the heart, it would update the number of "hearts" for this article somewhere in the database.
But actually, instead of updating the database... well... it does nothing, and returns a new, random number of hearts. Lame!

Look in the public/js directory: open article_show.js:

16 lines  public/js/article_show.js
$(document).ready(function() {
    $('.js-like-article').on('click', function(e) {
        e.preventDefault();
        var $link = $(e.currentTarget);
        $link.toggleClass('fa-heart-o').toggleClass('fa-heart');
        $.ajax({
            method: 'POST',
            url: $link.attr('href')
        }).done(function(data) {
            $('.js-like-article-count').html(data.hearts);
        })
    });
});
In that tutorial, we wrote some simple JavaScript that said: when the "like" link is clicked, toggle the styling on the heart, and then send a POST request to the URL that's in the href of the link. Then, when the AJAX call finishes, read the new number of hearts from the JSON response and update the page.

The href that we're reading lives in show.html.twig. Here it is:

 86 lines  templates/article/show.html.twig
{% block body %}
<div class="container">
    <div class="row">
        <div class="col-sm-12">
            <div class="show-article-container p-3 mt-4">
                <div class="row">
                    <div class="col-sm-12">
                        <div class="show-article-title-container d-inline-block pl-3 align-middle">
                            <span class="pl-2 article-details">
                                <a href="{{ path('article_toggle_heart', {slug: article.slug}) }}" class="fa fa-heart-o like-article js-like-article"></a>
                            </span>
                        </div>
                    </div>
                </div>
            </div>
        </div>
    </div>
</div>
{% endblock %}
It's a URL to some route called article_toggle_heart. And we're sending the article slug to that endpoint.
Open up ArticleController, and scroll down to find that route: it's toggleArticleHeart():

 74 lines  src/Controller/ArticleController.php
class ArticleController extends AbstractController
{
    /**
     * @Route("/news/{slug}/heart", name="article_toggle_heart", methods={"POST"})
     */
    public function toggleArticleHeart($slug, LoggerInterface $logger)
    {
        // TODO - actually heart/unheart the article!
        $logger->info('Article is being hearted! ');
        return new JsonResponse(['hearts' => rand(5, 100)]);
    }
}
And, as you can see... this endpoint doesn't actually do anything! Other than return JSON with a random number, which our JavaScript uses to update the page:

 16 lines  public/js/article_show.js
$(document).ready(function() {
    $('.js-like-article').on('click', function(e) {
        $.ajax({
        }).done(function(data) {
            $('.js-like-article-count').html(data.hearts);
        })
    });
});