13.2.Updating an Entity{Updating the heartCount}
Updating the heartCount
It's time to implement this feature correctly! Or, at least, more correctly. And, for the first time, we will update an existing row in the database.
Back in ArticleController, we need to use the slug to query for the Article object. But, remember, there's a shortcut for this: replace the $slug argument with Article $article
Thanks to the type-hint, Symfony will automatically try to find an Article with this slug.
Then, to update the heartCount, just $article->setHeartCount() and then $article->getHeartCount() + 1
Side note, it's not important for this tutorial, but in a high-traffic system, this could introduce a race condition. Between the time this article is queried for, and when it saves, 10 other people might have also liked the article. And that would mean that this would actually save the old, wrong number, effectively removing the 10 hearts that occurred during those microseconds.
Anyways, at the bottom, instead of the random number, use $article->getHeartCount().
So, now, to the key question: how do we run an UPDATE query in the database? Actually, it's the exact same as inserting a new article. Fetch the entity manager like normal: EntityManagerInterface $em.
Then, after updating the object, just call $em->flush().
But wait! I did not call $em->persist(). We could call this... it's just not needed for updates! When you query Doctrine for an object, it already knows that you want that object to be saved to the database when you call flush(). Doctrine is also smart enough to know that it should update the object, instead of inserting a new one.
Ok, go back and refresh! Here is the real heart count for this article: 88. Click the heart and... yea! 89! And if you refresh, it stays! We can do 90, 91, 92, 93, and forever! And yea... this is not quite realistic yet. On a real site, I should only be able to like this article one time. But, we'll need to talk about users and security before we can do that.