13.3.Updating an Entity{Smarter Entity Method}
Smarter Entity Method
Now that this is working, we can improve it! In the controller, we wrote some code to increment the heart count by one.
But, whenever possible, it's better to move code out of your controller. Usually we do this by creating a new service class and putting the logic there. But, if the logic is simple, it can sometimes live inside your entity class. Check this out: open Article, scroll to the bottom, and add a new method: public function incrementHeartCount(). Give it no arguments and return self, like our other methods.
Then, $this->heartCount = $this->heartCount + 1
Back in ArticleController, we can simplify to $article->incrementHeartCount()
That's so nice. This moves the logic to a better place, and, it reads really well:
Hello Article: I would like you to increment your heart count. Thanks!

Smart Versus Anemic Entities
And... this touches on a somewhat controversial topic related to entities. Notice that every property in the entity has a getter and setter method. This makes our entity super flexible: you can get or set any field you need.
But, sometimes, you might not need, or even want a getter or setter method. For example, do we really want a setHeartCount() method?
I mean, should any part of the app ever need to change this? Probably not: they should just call our more descriptive incrementHeartCount() instead.
I am going to keep it, because we use it to generate our fake data, but I want you to really think about this point.
By removing unnecessary getter or setter methods, and replacing them with more descriptive methods that fit your business logic, you can, little-by-little, give your entities more clarity. Some people take this to an extreme and have almost zero getters and setters. Here at KnpU, we tend to be more pragmatic: we usually have getters and setters, but we always look for ways to be more descriptive.
Next, our dummy article data is boring, and we're creating it in a hacky way.
Let's build an awesome fixtures system instead.