14.3.Fixtures: Seeding Dummy Data! {Writing the Fixtures}
Writing the Fixtures
Open ArticleAdminController: let's start stealing some code! Copy all of our dummy article code, go back to the fixture class, and paste! We need to re-type the e on Article and hit tab so that PhpStorm adds the use statement for us on top
Then, at the bottom, the entity manager variable is called $manager. Back in the controller, just put a die('todo') for now.
Someday, we'll create a proper admin form here.
And... that's it! It's super boring and it only creates one article... but it should work! Try it: find your terminal and run a new console command:

php bin/console doctrine:fixtures:load
This will ask if you want to continue because - important note! - the command will empty the database first, and then load fresh data. Again, not something you want to run on production... not saying I've done that before.
When it finishes, find your browser, and refresh. It works!
But... come on! We're going to need more than one article! How can we create multiple?