14.4.Fixtures: Seeding Dummy Data! {Creating Multiple Articles}
Creating Multiple Articles
First, we're going to do it the easy... but, kinda boring way. A for loop! Say: for  = 0;  < 10; ++. And, all the way at the bottom, add the ending curly brace. We need to call persist() in the loop, but we only need to call flush() once at the end.

Cool! Try it again:

php bin/console doctrine:fixtures:load
Then, refresh! Awesome! Except that the articles are still totally boring and identical... we'll talk about that in the next chapter.