14.5.Fixtures: Seeding Dummy Data! {BaseFixture Class for Cooler Looping}
BaseFixture Class for Cooler Looping
But first, let me show you a cooler way to create multiple articles. In the DataFixtures directory, create a new class called BaseFixture. Make it abstract, and extend the normal class that all fixtures extend... so... Fixture.
Here's the idea: this will not be a fixture class that the bundle will execute. Instead, it will be a base class with some cool helper methods. To start, copy the load() method and implement it here. Re-type ObjectManager to get its use statement.
Oh, ObjectManager is an interface implemented by EntityManager, it's not too important: just think "this is the entity manager".
Next, and this won't make sense yet, create a private  property, and set it inside the load() method.
Finally, create an abstract protected function called loadData() with that same ObjectManager argument. Back in load(), call this: $this->loadData($manager).
So far, this doesn't do anything special. Back in ArticleFixtures, extend the new BaseFixture instead. I'll also cleanup the extra use statement.
Now, instead of implementing load(), implement loadData() and make it protected.
And... yea! The fixture system will call load() on BaseFixture, that will call loadData() on ArticleFixtures and... well... everything will work exactly like before.