14.6.Fixtures: Seeding Dummy Data! {Adding the createMany Method}
Adding the createMany Method
So... why did we just do all the stuff from previous commit? Go back to the BaseFixture class and, at the bottom, I'm going to paste in a little method that I created.
Oh, and to make PhpStorm happy, at the top, add some PHPDoc that the $manager property is an ObjectManager instance.
Anyways, say hello to createMany()! A simple method that we can call to create multiple instances of an object. Here's the idea: we call createMany() and pass it the class we want to create, how many we want to create, and a callback method that will be called each time an object is created. That'll be our chance to load that object with data.
Basically, it does the for loop for us... which is not a huge deal, except for two nice things. First, it calls persist() for us, so we don't have to.
Ok, cool, but not amazing. But, this last line is cool:
abstract class BaseFixture extends Fixture
{
    protected function createMany(string $className, int $count, callable $factory)
    {
        for ($i = 0; $i < ; $i++) {
            // store for usage later as App\Entity\ClassName_#COUNT#
            $this->addReference($className . '_' . $i, $entity);
        }
    }
}
It won't matter yet, but in a future tutorial, we will have multiple fixtures classes. When we do, we will need to be able to reference objects created in one fixture class from other fixture classes. By calling addReference(), all of our objects are automatically stored and ready to be fetched with a key that's their class name plus the index number.

The point is: this is going to save us some serious work... but not until the next tutorial.