14.7.Fixtures: Seeding Dummy Data! {Adding the createMany Method}
Back in ArticleFixtures, use the new method: $this->createMany() passing it Article::class, 10, and a function.
This will have two args: the Article that was just created and a count of which article this is. Inside the method, we can remove the $article = new Article(), and instead of a random number on the slug, we can use $count.
At the bottom, the persist isn't hurting anything, but it's not needed anymore.
Finish the end with a closing parenthesis and a semicolon.
So, it's a little bit fancier, and it'll save that important reference for us. Let's try it! Reload the fixtures again:

php bin/console doctrine:fixtures:load
No errors! Refresh the homepage: ah, our same, boring list of 10 identical articles. In the next chapter, let's use an awesome library called Faker to give each article rich, unique, realistic data.