17.4.When Migrations Fail{When a Migration Fails}
To finally make the fields not nullable, we can ask Doctrine to generate a new migration:

php bin/console make:migration
Go check it out!

Ha! Nice! It simply changes the fields to be NOT NULL. Run it!

php bin/console doctrine:migrations:migrate
And we are good! Now, back to Timestampable!